requirements.txt
alembic==1.2.0
falcon==2.0.0
falcon-auth==1.1.0
falcors==2.0.0
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.0
marshmallow-sqlalchemy==0.19.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
# simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.8
ujson==1.35
